ds-ai-pipeline/src/requirements.arm.txt
azure-iot-device
opentelemetry-sdk==1.13.0
azure-monitor-opentelemetry-exporter==1.0.0b8
furl
scipy
numpy
pyyaml